backend/requirements.txt
# Core FastAPI stack
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.40
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0  # For async PostgreSQL support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1

# Pydantic
pydantic==2.10.3
pydantic-settings==2.6.0
email-validator==2.2.0

# Async & Performance
httpx==0.28.1
asyncio-throttle==1.0.2
aiosqlite==0.20.0

# Monitoring & Logging
structlog==24.4.0
loguru==0.7.2
prometheus-fastapi-instrumentator==7.1.0

# Caching & Rate Limiting
redis==5.2.0
aioredis==2.0.1
slowapi==0.1.9

# Validation & Serialization
pydantic-extra-types==2.6.0
pydantic[email]==2.10.3

# Development
black==24.10.0
isort==5.13.2
mypy==1.13.0
ruff==0.9.4
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==5.0.0